Deliver context menu event with correct coordinates to widgets in popups
If a popup is open, Qt delivers events to the popup child under the
mouse, so we need to correctly translate the local position of the
context menu event to that child's coordate system.

This is already done correctly for regular mouse events, so use the same
logic here.

Fixes: QTBUG-99371
Change-Id: Ief24c755e76d4d1aa2304b06662ed26ae309f684
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 0462d24dcd851222d4ac73e7af38d785c6a32cfd)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>