Use IS_SYMLINK condition instead of REALPATH when checking BINARY_DIR
For case-insensitive file systems, the REALPATH file component may
return file paths in a different case than ABSOULEPATH. We use REALPATH
to detect symlinks. In case-sensitive file systems there still might be
the situation when /path/to/Build is a symlink to the /path/to/build,
so we cannot make the comparison case-insensitive. Use the IS_SYMLINK
check instead.

Pick-to: 6.2 6.3
Fixes: QTBUG-99416
Change-Id: Idf998b9e09837a834139553fb822e05b2a0a1949
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>