QWindowPrivate: de-inline dtor and ctor
It's private API, but exported, so de-inline the dtor to pin the
vtable in QtGui instead of potentially duplicating it in every library
that uses the class.

Ditto ctor, but that's just code hygiene: we don't want the code to be
duplicated across all users.

Task-number: QTBUG-45582
Change-Id: I91ea38be20fc67795466a68ca5721837255b33a0
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit b31c4e38dc7015663825ee46d124a8759388d67f)
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>